backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
